requirements.txt
paramiko==3.4.0
python-docx==1.1.0
python-dotenv==1.0.0
schedule==1.2.0
croniter==2.0.1
docx2txt==0.9
pytz==2023.3

# Note: ftplib (for FTPS support) is part of Python standard library
# Note: concurrent.futures (for parallel execution) is part of Python standard library
